Revamped caching.  FishService now has a Level 1 (memory) and Level 2 (memcache) cache.  FishesServlet has a Level 1 (memory) cache for the served JSON effectively creating a cache for the serialized version of the data.  Cache invalidation now also works correctly.
Overall latency should improve - even immediately after a server push
due to the Level 2 cache in FishService.

+++ src/ffxiv/fisher/service/FishService.java
@@ -5,34 +5,72 @@ import static com.google.common.base.Preconditions.checkArgument;
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
+import java.util.concurrent.atomic.AtomicReference;
+import java.util.logging.Logger;
 
 import javax.annotation.Nullable;
 
 import com.google.appengine.api.datastore.Key;
 import com.google.appengine.api.datastore.KeyFactory;
+import com.google.appengine.api.memcache.MemcacheService;
 import com.google.code.twig.ObjectDatastore;
 import com.google.inject.Inject;
 import com.google.inject.Provider;
 import com.google.inject.Singleton;
 
+import ffxiv.fisher.Annotations.DevelopmentEnvironment;
 import ffxiv.fisher.model.Fish;
+import ffxiv.fisher.model.FishSerializer;
 
 /**
  * Service for interacting with fish data.
  */
 @Singleton
 public class FishService {
-	
+
 	/**
-	 * Memcache key for the JSON version of the fish.
+	 * Allows other layers to invalidate their caches when fish change.
 	 */
-	public static final String FISHES_KEY = "fishes";
+	public static interface InvalidationCallback {
+		void invalidate();
+	}
 	
+	private static final Logger log = Logger.getLogger(FishService.class.getName());
+	
+	private static final String MEMCACHE_FISHES_KEY = "fishes";
+	
+	// Level 1 cache: store the fish in memory.
+	// Fastest when available (nearly instant).
+	private final AtomicReference<List<Fish>> cachedFishRef;
+	
+	// Level 2 cache: memcache.  Helps when pushing a new server.
+	// Second fastest when available (~20-40ms).
+	private final MemcacheService memcacheService;
+	
+	// The source of truth for fish.  Sloooooooow (on the order of seconds).
 	private final Provider<ObjectDatastore> datastoreProvider;
 	
+	private final FishSerializer fishSerializer;
+	private final List<InvalidationCallback> invalidationCallbacks;
+	private final boolean isDevelopmentEnvironment;
+	
 	@Inject
-	public FishService(Provider<ObjectDatastore> datastoreProvider) {
+	public FishService(
+			Provider<ObjectDatastore> datastoreProvider,
+			MemcacheService memcacheService,
+			FishSerializer fishSerializer,
+			@DevelopmentEnvironment boolean isDevelopmentEnvironment) {
 		this.datastoreProvider = datastoreProvider;
+		this.memcacheService = memcacheService;
+		this.fishSerializer = fishSerializer;
+		this.isDevelopmentEnvironment = isDevelopmentEnvironment;
+
+		cachedFishRef = new AtomicReference<>();
+		invalidationCallbacks = new ArrayList<>();
+	}
+	
+	public void registerInvalidationCallback(InvalidationCallback callback) {
+		invalidationCallbacks.add(callback);
 	}
 	
 	/**
@@ -40,6 +78,7 @@ public class FishService {
 	 */
 	@Nullable
 	public Fish get(Key key) {
+		// TODO Read from Level 1 and 2 caches first.
 		ObjectDatastore datastore = datastoreProvider.get();
 		Fish fish = datastore.load(key);
 		if (fish != null) {
@@ -52,6 +91,25 @@ public class FishService {
 	 * Gets all fish.
 	 */
 	public List<Fish> getAll() {
+		// Read from Level 1 cache first.  Should happen frequently.
+		List<Fish> cachedFish = cachedFishRef.get();
+		if (cachedFish != null) {
+			log.info("Returning fish from memory.");
+			return cachedFish;
+		}
+		
+		// Read from Level 2 cache next.  Should only happen after a server push.
+		String cachedFishJson = (String) memcacheService.get(MEMCACHE_FISHES_KEY);
+		if (cachedFishJson != null) {
+			cachedFish = fishSerializer.deserializeAll(cachedFishJson);
+			// Update Level 1 cache from memcache.
+			cachedFishRef.compareAndSet(null, cachedFish);
+			log.info("Returning fish from memcache.");
+			return cachedFish;
+		}
+		
+		// Only read from the database if we must.  Shouldn't really happen unless
+		// data was written and caches were invalidated.
 		ObjectDatastore datastore = datastoreProvider.get();
 		Iterator<Fish> iterator = datastore
 				.find()
@@ -65,10 +123,17 @@ public class FishService {
 			fish.setKey(KeyFactory.keyToString(key));
 			fishList.add(fish);
 		}
+		
+		// Update both Level 1 and Level 2 caches.
+		cachedFishRef.compareAndSet(null, fishList);
+		memcacheService.put(MEMCACHE_FISHES_KEY, fishSerializer.serialize(fishList));
+		
+		log.info("Returning fish from the database.");
 		return fishList;
 	}
 	
 	public Fish get(String key) {
+		// TODO(jdeyerle): Utilize cached fish.
 		ObjectDatastore datastore = datastoreProvider.get();
 		Fish fish = datastore.load(KeyFactory.stringToKey(key));
 		fish.setKey(key);
@@ -87,6 +152,10 @@ public class FishService {
 		loadedFish.setFromFish(fish);
 		datastore.update(loadedFish);
 		
+		// TODO Be smarter and just update the cache instead.  This operation
+		// is very infrequent, though.
+		invalidateCaches();
+		
 		return loadedFish;
 	}
 	
@@ -104,6 +173,10 @@ public class FishService {
 		return create(fish, false);
 	}
 	
+	/**
+	 * Creates a fish but with the option of skipping validation.  Skipping
+	 * validation is only used in the admin path.
+	 */
 	private Fish create(Fish fish, boolean withValidation) {
 		if (withValidation) {
 			validate(fish, true);
@@ -115,8 +188,17 @@ public class FishService {
 		Key key = datastore.associatedKey(fish);
 		fish.setKey(KeyFactory.keyToString(key));
 		
+		// TODO Be smarter and just update the cache instead.  This operation
+		// is very infrequent, though.
+		invalidateCaches();
+		
 		return fish;
 	}
+	
+	/**
+	 * Validates the fish in question.  The path is shared between update and create.
+	 * An {@link IllegalArgumentException} is thrown if anything is invalid.
+	 */
 	private void validate(Fish fish, boolean create) throws IllegalArgumentException {
 		if (create) {
 			checkArgument(fish.getKey() == null || fish.getKey().isEmpty(),
@@ -135,8 +217,52 @@ public class FishService {
 				"End hour must be between 0 and 24 inclusive.");
 	}
 	
+	/**
+	 * Deletes all the fish in the database.  This is only called from the development
+	 * environment when refreshing the database.  This should NEVER be called in
+	 * production.
+	 */
 	public void deleteAll() {
+		// Safeguard against deleting production data.
+		if (!isDevelopmentEnvironment) {
+			log.severe("Attempted to delete all fish in production.");
+			return;
+		}
+		
 		ObjectDatastore datastore = datastoreProvider.get();
 		datastore.deleteAll(Fish.class);
+
+		// Invalidate caches on write.
+		invalidateCaches();
+	}
+	
+	/**
+	 * This is a hack.  It is used to try and keep memcache data around by
+	 * assuming that anything placed in memcache has a TTL.  By reading memcache
+	 * we are assuming it's refreshed and kept alive.
+	 */
+	public void keepMemcacheAlive() {
+		log.info("Keeping memcache alive");
+		Object allFishJson = memcacheService.get(MEMCACHE_FISHES_KEY);
+		if (allFishJson != null) {
+			log.info("Memcache is still fresh");
+		} else {
+			log.info("Memcache has no fish");
+		}
+	}
+	
+	/**
+	 * Invalidates all caches for fish.  Used whenever a large mutation happens.
+	 */
+	private void invalidateCaches() {
+		log.info("Invalidating all fish caches.");
+		
+		cachedFishRef.set(null);
+		memcacheService.delete(MEMCACHE_FISHES_KEY);
+		
+		// Notify all registered callbacks of invalidation.
+		for (InvalidationCallback callback : invalidationCallbacks) {
+			callback.invalidate();
+		}
 	}
 }
